requirements.txt
flask
flask-socketio
opencv-python
pyyaml